+++ src/App.tsx
@@ -1,4 +1,3 @@
-import "./App.scss";
 import { lazy } from "react";
 import { Route, Routes } from "react-router-dom";
