+++ src/common/types/navbar/index.ts
@@ -1,4 +1,4 @@
-export interface ITopBarProps {
+export interface INavBarProps {
   isOpen: boolean;
   setIsOpen: (value: boolean) => void;
   isNonMobile: boolean;
