+++ src/common/types/sidebar/index.ts
@@ -1,0 +1,6 @@
+export interface ISidebarProps {
+  isNonMobile: boolean;
+  drawerWidth: string;
+  isOpen: boolean;
+  setIsOpen: (value: boolean) => void;
+}
