+++ src/components/AppButton/AppButton.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { styled } from "@mui/styles";
+import { Button } from "@mui/material";
+
+const AppButton = styled(Button)({
+  borderRadius: 4,
+  backgroundColor: "#1900D5 !important",
+  boxShadow: "0px 1px 7px #332a76 !important",
+  padding: "10px 20px !important",
+  color: "white !important",
+  maxWidth: 300,
+});
+
+export default AppButton;
