--- src/components/Home/Home.scss
@@ -1,6 +1,0 @@
-.home {
-    min-height: calc(100vh - 110px);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
