+++ src/components/Home/styles.ts
@@ -1,0 +1,10 @@
+import { makeStyles } from "@mui/styles";
+
+export const useStyles = makeStyles({
+  home: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    minHeight: "calc(100vh - 110px)",
+  },
+});
