+++ src/components/auth/styles.ts
@@ -1,0 +1,15 @@
+import { makeStyles } from "@mui/styles";
+
+export const useStyles = makeStyles({
+  root: {
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    width: "100vw",
+    height: "100vh",
+    padding: "20px",
+  },
+  form: {
+    flex: 1,
+  },
+});
