+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.scss';
+import './index.css';
 import { App } from './App';
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
