--- src/pages/AuthPage/AuthPage.scss
@@ -1,19 +1,0 @@
-@import "../../assets/utils/variables.scss";
-
-
-.auth-root {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-
-    .auth-form {
-        width: 80%;
-
-        .incitingText {
-            color: $blue;
-            margin-left: 10px;
-            cursor: pointer;
-        }
-    }
-}
